DB 연동을 위해 다시 코드 되돌리고 재작성, 회원가입 기능 구현
