+++ package.json
@@ -14,6 +14,8 @@
     "crypto": "^1.0.1",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
+    "google-auth-library": "^9.11.0",
+    "googleapis": "^140.0.0",
     "mysql2": "^3.10.1"
   },
   "devDependencies": {
